+++ R/format-bytes.R
@@ -1,14 +1,14 @@
 #' Bytes formatter: convert to byte measurement and display symbol
 #'
-#' @return a function with three parameters, `x`, a numeric vector that returns
-#'   a character vector, `symbol` the byte symbol (e.g. "kB") desired and the
-#'   measurement `units` (traditional `binary` or `si` for SI metric units).
-#' @param x a numeric vector to format
+#' @inheritParams number_format
 #' @param symbol byte symbol to use. If "auto" the symbol used will be
 #'   determined separately for each value of `x`. Valid symbols are "B", "kB",
 #'   "MB", "GB", "TB", "PB", "EB", "ZB", and "YB" for SI units, and the "iB"
 #'   variants for binary units.
 #' @param units which unit base to use, "binary" (1024 base) or "si" (1000 base)
+#' @return a function with three parameters, `x`, a numeric vector that returns
+#'   a character vector, `symbol` the byte symbol (e.g. "kB") desired and the
+#'   measurement `units` (traditional `binary` or `si` for SI metric units).
 #' @param ... other arguments passed to [number()]
 #' @references Units of Information (Wikipedia) :
 #'   \url{http://en.wikipedia.org/wiki/Units_of_information}
@@ -26,7 +26,7 @@ number_bytes_format <- function(symbol = "auto", units = "binary", ...) {
 
 #' @export
 #' @rdname number_bytes_format
-number_bytes <- function(x, symbol = "auto", units = c("binary", "si"), ...) {
+number_bytes <- function(x, symbol = "auto", units = c("binary", "si"), accuracy = 1, ...) {
   units <- match.arg(units, c("binary", "si"))
 
   powers <- si_powers[si_powers >= 3] / 3 # powers of 1000
@@ -47,7 +47,7 @@ number_bytes <- function(x, symbol = "auto", units = c("binary", "si"), ...) {
     power <- match(symbol, symbols) - 1L
   }
 
-  number(x / base^power, suffix = paste0(" ", symbol), ...)
+  number(x / base^power, accuracy = accuracy, suffix = paste0(" ", symbol), ...)
 }
 
 validate_byte_symbol <- function(symbol, symbols, default = "auto") {