+++ tests/testthat/test-format-bytes.R
@@ -4,7 +4,7 @@ test_that("Byte formatter can take a symbol designator", {
     c("50 B", "400 B", "502 B", NA)
   )
   expect_equal(
-    number_bytes(c(3e6, 4e6, 5e6), symbol = "MiB"),
+    number_bytes(3:5 * 1024^2, symbol = "MiB"),
     c("3 MiB", "4 MiB", "5 MiB")
   )
   expect_equal(