+++ tests/testthat/test-format-number.r
@@ -70,7 +70,7 @@ test_that("comma preserves names", {
 test_that("unit format", {
   expect_equal(
     unit_format(unit = "km", scale = 1e-3)(c(1e3, NA, 2e3)),
-    c("1 km", NA, "2 km")
+    c("1.0 km", NA, "2.0 km")
   )
   expect_equal(
     unit_format(unit = "ha", scale = 1e-4, accuracy = .1)(c(1e3, 2e3)),
@@ -123,8 +123,8 @@ test_that("preserves names", {
 
 test_that("default accuracy works for range of inputs", {
   x <- c(0.1, 0.2, 0.5)
-  expect_equal(percent(x / 100), c("0.1%", "0.2%", "0.5%"))
-  expect_equal(percent(x / 10), c("1%", "2%", "5%"))
+  expect_equal(percent(x / 100), c("0.10%", "0.20%", "0.50%"))
+  expect_equal(percent(x / 10), c("1.0%", "2.0%", "5.0%"))
   expect_equal(percent(x), c("10%", "20%", "50%"))
   expect_equal(percent(x * 10), c("100%", "200%", "500%"))
 })